+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 -   Add support for managing side effects
 -   Make subviews, delegated events and side effects storage "private"
     properties
+-   Add support for delegated `focus` and `blur` events
 -   Document parent/child communication
 
 ## [6.0.1][] - 2019-04-04